Create carddeck-logback-config.yaml
In this YAML file, a ConfigMap named carddeck-webapp-logback-config is mounted as a volume at the path /usr/local/tomcat/conf/logback.xml, and the logback-config.yaml file is referenced as logback.xml. The subPath field is used to specify the file name in the volume mount path. Once the Deployment is deployed, the logs for the Carddeck-webapp should be written to the standard output stream and visible in the K9S console (personally I prefer to use it :)).